+++ src/jsymbolic2/features/VerticalIntervalHistogramFeature.java
@@ -67,7 +67,7 @@ public class VerticalIntervalHistogramFeature extends MIDIFeatureExtractor {
             throws Exception
     {
         // Get the vertical interval intermediate representation chart
-        int[][] vertical_interval_chart = sequence_info.vertical_interval_chart;
+        short[][] vertical_interval_chart = sequence_info.vertical_interval_chart;
 
         // An array for all possible interval values
         double[] all_intervals = new double[128];
