+++ src/jsymbolic2/features/VerticalIntervalWrappedHistogramFeature.java
@@ -35,7 +35,7 @@ public class VerticalIntervalWrappedHistogramFeature extends MIDIFeatureExtracto
             throws Exception
     {
         // Get the vertical interval intermediate representation chart
-        int[][] vertical_interval_chart = sequence_info.vertical_interval_chart;
+        short[][] vertical_interval_chart = sequence_info.vertical_interval_chart;
 
         // An array for all possible interval values
         double[] all_intervals = new double[12];
